Divided js over more files
